+++ tests/devices/default_tensor/test_default_tensor.py
@@ -25,7 +25,7 @@ from scipy.sparse import csr_matrix
 
 import pennylane as qml
 from pennylane.exceptions import DeviceError
-from pennylane.qchem import givens_decomposition
+from pennylane.math.decomposition import givens_decomposition
 from pennylane.typing import TensorLike
 from pennylane.wires import WireError
